Comment view controller refactoring and added documentation. Also, minor fixes.
