+++ Slide for Reddit/PagingCommentViewController.swift
@@ -5,7 +5,6 @@
 //  Created by Carlos Crane on 7/23/17.
 //  Copyright © 2017 Haptic Apps. All rights reserved.
 //
-
 import Foundation
 
 class PagingCommentViewController: ColorMuxPagingViewController, UIPageViewControllerDataSource, UIPageViewControllerDelegate {
@@ -42,7 +41,7 @@ class PagingCommentViewController: ColorMuxPagingViewController, UIPageViewContr
                                    direction: .forward,
                                    animated: true,
                                    completion: nil)
-            vc.refresh(vc)
+            vc.refreshComments(vc)
         }
     }
     
@@ -142,7 +141,7 @@ class PagingCommentViewController: ColorMuxPagingViewController, UIPageViewContr
         if !(self.viewControllers!.first! is ClearVC) {
             PagingCommentViewController.savedComment = self.viewControllers!.first as? CommentViewController
             if !PagingCommentViewController.savedComment!.loaded {
-                PagingCommentViewController.savedComment!.refresh(pageViewController)
+                PagingCommentViewController.savedComment!.refreshComments(pageViewController)
             }
 
         }
@@ -200,7 +199,7 @@ class PagingCommentViewController: ColorMuxPagingViewController, UIPageViewContr
         }
 
         if !(viewController as! CommentViewController).loaded {
-            (viewController as! CommentViewController).refresh(viewController)
+            (viewController as! CommentViewController).refreshComments(viewController)
         }
         
         let nextIndex = viewControllerIndex + 1
